Update user-groups.php to avoid resaving usergroup for user already in group
This represents an enormous performance improvement when there are a lot of users and groups. 

Without this check if the user is already in the group it gets re-saved anyway. If the user is in many groups this represents a lot of wasted time, and on our site it was slowing down user saving a LOT. 

I see no risk in this check being added to avoid re-saving a user who is already in a group. 

Thank you for considering it.