Calendar: Remove 'taxonomy hierarchical' from editable fields for performance and usability
Re: https://github.com/Automattic/Edit-Flow/issues/756

Stop hierarchical taxonomies like categories from being edited inside the calendar, while continuing to allow non-hierarchical taxonomies (tags) to be edited.

This avoids the horrific performance punishment on sites with many categories due to the full list of categories being inserted as a hidden field inside each and every post on the calendar.

It also avoids using a single-select field to update categories, which removes all categories but the one chosen when saved. Users are better off editing the post directly if they need to change the categories, at least unless and until we can offer a proper multi-select category picker in the calendar (a terrible idea because of the performance issues mentioned above).